Multiple enhancements, fixes and changes (see details)
# Replaced all snow and doParallel functionality with future package. The cores
argument is now removed from options of all functions and instead the user
can just setup a plan via the future package which is much simpler and allows
much more customization. This may break backwards compatibility with versions
less than this which made use of the cores argument.
# Bumped version to 0.3.0 across all affected packages.
# Made solver nlminb and autodiff (TRUE) defaults across all functions.
# Removed options for drop_na, drop_negative and redraw in the prediction
function. The user is left to choose what to do when a model is mispecified
so badly that it returns negative values. The problem with removing negatives
is that it biases the mean considerably effectively imposting a half normal
distribution (not an assumption of the estimation for additive models).
# Added analytic moments (tsmoments) method for class of models 1, 2, 3 and
approximate function (exact for h<m) for class 3 with regressors.
# Fixed a bug in the bounds of the multiplicative error model for the seaasonal
initialization when estimate was set to TRUE.
# The class of model is now returned in the specification object
# Next version will look into adding admissible bounds (code from R.Hyndman),
and fixing how the seed value gets passed.